Update neuralert.c
This patch solves a hypothesized cause of a hard fault seen in one device after 3500 successful transmissions.

The reasoning is described in the code, but the assumption is that the LmacMain task needs to cleanly complete before the rf is turned off -- the delay in the existing code (10 ticks) was insufficient to block the rf if an accelerometer interrupt was called which is the highest priority application task.  To resolve this, we observed that the accelerometer interrupt task is about 250 ms, so we added a 1 second delay between when the wifi is disconnected and when the rf is turned off -- giving plenty of time for the LmacMain task to shut down.